SDIT-3017: ♻️ Combine search clients (#336)
