+++ server/services/index.ts
@@ -8,21 +8,15 @@ import AgencySearchService from './agencySearchService'
 import MigratePrisonerService from './migratePrisonerService'
 
 export const services = () => {
-  const {
-    applicationInfo,
-    manageUsersApiClient,
-    prisonApiClient,
-    prisonerSearchClient,
-    restrictedPatientApiClient,
-    restrictedPatientSearchClient,
-  } = dataAccess()
+  const { applicationInfo, manageUsersApiClient, prisonApiClient, prisonerSearchClient, restrictedPatientApiClient } =
+    dataAccess()
 
   return {
     applicationInfo,
     userService: new UserService(manageUsersApiClient),
     prisonerSearchService: new PrisonerSearchService(prisonApiClient, prisonerSearchClient),
     movePrisonerService: new MovePrisonerService(restrictedPatientApiClient),
-    restrictedPatientSearchService: new RestrictedPatientSearchService(prisonApiClient, restrictedPatientSearchClient),
+    restrictedPatientSearchService: new RestrictedPatientSearchService(prisonApiClient, prisonerSearchClient),
     removeRestrictedPatientService: new RemoveRestrictedPatientService(prisonApiClient, restrictedPatientApiClient),
     agencySearchService: new AgencySearchService(prisonApiClient),
     migratePrisonerService: new MigratePrisonerService(restrictedPatientApiClient),
