+++ server/services/restrictedPatientSearchService.test.ts
@@ -1,13 +1,13 @@
 import 'reflect-metadata'
 
 import RestrictedPatientSearchService from './restrictedPatientSearchService'
-import RestrictedPatientSearchClient from '../data/restrictedPatientSearchClient'
+import PrisonerSearchClient from '../data/prisonerSearchClient'
 import RestrictedPatientSearchResult from '../data/restrictedPatientSearchResult'
 import PrisonApiClient, { Agency } from '../data/prisonApiClient'
 
 import { Context } from './context'
 
-jest.mock('../data/restrictedPatientSearchClient')
+jest.mock('../data/prisonerSearchClient')
 jest.mock('../data/prisonApiClient')
 
 const user = {
@@ -15,13 +15,13 @@ const user = {
 } as Context
 
 const prisonApiClient = new PrisonApiClient(null) as jest.Mocked<PrisonApiClient>
-const restrictedPatientSearchClient = new RestrictedPatientSearchClient() as jest.Mocked<RestrictedPatientSearchClient>
+const prisonerSearchClient = new PrisonerSearchClient(null) as jest.Mocked<PrisonerSearchClient>
 
 describe('restrictedPatientSearchService', () => {
   let service: RestrictedPatientSearchService
 
   beforeEach(() => {
-    service = new RestrictedPatientSearchService(prisonApiClient, restrictedPatientSearchClient)
+    service = new RestrictedPatientSearchService(prisonApiClient, prisonerSearchClient)
     prisonApiClient.getAgenciesByType.mockResolvedValue([
       {
         agencyId: 'MDI',
@@ -46,7 +46,7 @@ describe('restrictedPatientSearchService', () => {
 
   describe('search', () => {
     it('search by prisoner identifier', async () => {
-      restrictedPatientSearchClient.search.mockResolvedValue([
+      prisonerSearchClient.restrictedPatientSearch.mockResolvedValue([
         {
           alerts: [
             { alertType: 'T', alertCode: 'TCPA' },
@@ -76,11 +76,11 @@ describe('restrictedPatientSearchService', () => {
         ]),
       )
 
-      expect(restrictedPatientSearchClient.search).toBeCalledWith({ prisonerIdentifier: 'A1234AA' }, user.token)
+      expect(prisonerSearchClient.restrictedPatientSearch).toBeCalledWith({ prisonerIdentifier: 'A1234AA' }, user.token)
     })
 
     it('search by prisoner name', async () => {
-      restrictedPatientSearchClient.search.mockResolvedValue([
+      prisonerSearchClient.restrictedPatientSearch.mockResolvedValue([
         {
           firstName: 'JOHN',
           lastName: 'SMITH',
@@ -91,31 +91,40 @@ describe('restrictedPatientSearchService', () => {
 
       const results = await service.search({ searchTerm: 'Smith, John' }, user)
       expect(results).toEqual(expect.arrayContaining([expect.objectContaining({ displayName: 'Smith, John' })]))
-      expect(restrictedPatientSearchClient.search).toBeCalledWith({ lastName: 'Smith', firstName: 'John' }, user.token)
+      expect(prisonerSearchClient.restrictedPatientSearch).toBeCalledWith(
+        { lastName: 'Smith', firstName: 'John' },
+        user.token,
+      )
     })
 
     it('search by prisoner surname only', async () => {
       await service.search({ searchTerm: 'Smith' }, user)
-      expect(restrictedPatientSearchClient.search).toBeCalledWith({ lastName: 'Smith' }, user.token)
+      expect(prisonerSearchClient.restrictedPatientSearch).toBeCalledWith({ lastName: 'Smith' }, user.token)
     })
 
     it('search by prisoner name separated by a space', async () => {
       await service.search({ searchTerm: 'Smith John' }, user)
-      expect(restrictedPatientSearchClient.search).toBeCalledWith({ lastName: 'Smith', firstName: 'John' }, user.token)
+      expect(prisonerSearchClient.restrictedPatientSearch).toBeCalledWith(
+        { lastName: 'Smith', firstName: 'John' },
+        user.token,
+      )
     })
 
     it('search by prisoner name separated by many spaces', async () => {
       await service.search({ searchTerm: '    Smith   John ' }, user)
-      expect(restrictedPatientSearchClient.search).toBeCalledWith({ lastName: 'Smith', firstName: 'John' }, user.token)
+      expect(prisonerSearchClient.restrictedPatientSearch).toBeCalledWith(
+        { lastName: 'Smith', firstName: 'John' },
+        user.token,
+      )
     })
 
     it('search by prisoner identifier with extra spaces', async () => {
       await service.search({ searchTerm: '    A1234AA ' }, user)
-      expect(restrictedPatientSearchClient.search).toBeCalledWith({ prisonerIdentifier: 'A1234AA' }, user.token)
+      expect(prisonerSearchClient.restrictedPatientSearch).toBeCalledWith({ prisonerIdentifier: 'A1234AA' }, user.token)
     })
 
     it('augments supporting prison with description', async () => {
-      restrictedPatientSearchClient.search.mockResolvedValue([
+      prisonerSearchClient.restrictedPatientSearch.mockResolvedValue([
         {
           firstName: 'JOHN',
           lastName: 'SMITH',
